+++ UnitTest/TestSolver.cpp
@@ -24,7 +24,8 @@ TEST_CASE("GMRES Solver", "[Utility.Solver]") {
         vec x;
 
         SolverSetting<double> setting;
-        setting.preconditioner = std::make_unique<Jacobi>(A.A);
+        auto preconditioner = Jacobi<double>(A.A);
+        setting.preconditioner = &preconditioner;
 
         GMRES(&A, x, b, setting);
 
@@ -40,7 +41,8 @@ TEST_CASE("BiCGSTAB Solver", "[Utility.Solver]") {
         vec x;
 
         SolverSetting<double> setting;
-        setting.preconditioner = std::make_unique<Jacobi>(A.A);
+        auto preconditioner = Jacobi<double>(A.A);
+        setting.preconditioner = &preconditioner;
 
         BiCGSTAB(&A, x, b, setting);
 
@@ -64,7 +66,8 @@ TEST_CASE("Iterative Solver Sparse", "[Matrix.Solver]") {
         for(auto J = B.begin(); J != B.end(); ++J) A.at(J.row(), J.col()) = *J;
 
         SolverSetting<double> setting;
-        setting.preconditioner = std::make_unique<Jacobi>(A);
+        auto preconditioner = Jacobi<double>(A);
+        setting.preconditioner = &preconditioner;
 
         BiCGSTAB(&A, x, C, setting);
 
@@ -110,5 +113,30 @@ TEST_CASE("Iterative Solver Dense", "[Matrix.Solver]") {
         A.iterative_solve(x, C);
 
         REQUIRE(norm(solve(B, C) - x) <= 1E1 * setting.tolerance);
+    }
+}
+
+TEST_CASE("Iterative Solver Sparse Mat", "[Matrix.Solver]") {
+    SolverSetting<double> setting;
+    setting.iterative_solver = IterativeSolver::BICGSTAB;
+    setting.preconditioner_type = PreconditionerType::ILU;
+
+    for(auto I = 0; I < 10; ++I) {
+        const auto N = randi<uword>(distr_param(100, 200));
+        auto A = SparseMatSuperLU<double>(N, N);
+        REQUIRE(A.n_rows == N);
+        REQUIRE(A.n_cols == N);
+
+        sp_mat B = sprandu(N, N, .02) + speye(N, N) * 1E1;
+
+        const mat C = randu<mat>(N, 2 * N);
+        mat x;
+
+        A.zeros();
+        for(auto J = B.begin(); J != B.end(); ++J) A.at(J.row(), J.col()) = *J;
+        A.set_solver_setting(setting);
+        A.iterative_solve(x, C);
+
+        REQUIRE(norm(spsolve(B, C) - x) <= 1E-12);
     }
 }