+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 1. reformulate NM sections
 2. add couple stress membranes `CST3`, `CST6`, `CSM4-8`
-3. add universal iterative solvers `BiCGSTAB` and `GMRES`
+3. add universal iterative solvers `BiCGSTAB` and `GMRES`, and preconditioners `Jacobi` and `ILU`
 4. add support for `icx` and `ifx` compilers
 
 ## version 2.4